pyproject.toml
[tool.poetry]
name = "duplicate-reports"
version = "0.1.0"
description = ""
authors = ["WIMT"]
readme = "README.md"
packages = [{include = "duplicate_reports"}]

[tool.poetry.dependencies]
python = ">3.8"
scikit-learn = "^1.2.2"
blis = "^0.9.1"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.9.0"
google-cloud-bigquery = "^3.11.1"
openai = "^0.27.8"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
